Completed GetRecordData and fixed issues with position setting and reading in FileReader. Fixed small bug in integration test. Test still failing.

+++ Siftan/IRecordSource.cs
@@ -11,7 +11,7 @@ namespace Siftan
 
     void Close();
 
-    Int64 GetRecordData(Byte[] buffer);
+    Int32 GetRecordData(Byte[] buffer);
 
     Boolean MoveToNextRecord();
